Module contentextractor created
